Small tweaks to make fix

+++ PC/Seriel/SerialProtocol.cpp
@@ -96,7 +96,7 @@ bool SerialProtocol::stopRoutine() {
 }
 
 int SerialProtocol::sendChar(char data) {
-	std::cout << "Sending -" << data << "-" << std::endl; // Debugging purposes
+	std::cout << "Sending -" << (int)data << "-" << std::endl; // Debugging purposes
 	return cs.SendData(&data,1);
 }
 
@@ -110,6 +110,6 @@ char SerialProtocol::readData() {
 	char data = *lpBuffer;
 	delete lpBuffer;
 
-	cout << "Reading -" << data << "-" << endl; // Debugging purposes
+	cout << "Reading -" << (int)data << "-" << endl; // Debugging purposes
 	return data;
 }
